Sample
This was a project I built for TMC 498 Pro-Seminar at Arizona State University. The goal was to build an advertising model that didn't rely on tracking cookies yet would still give advertisers an idea of where their ads picked up traffic. This project was made possible with SMTPJS (www.smtpjs.com); it was used to help set up the email portion.